+++ ChangeLog.md
@@ -1,5 +1,13 @@
 # OldNews ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Using `NextUnread` and `PreviousUnread` while inside the navigation panel
+  now moves the highlight to the next folder or subscription that has unread
+  articles. ([#15](https://github.com/davep/oldnews/pull/15))
+
 ## v0.1.0
 
 **Released: 2026-01-16**